+++ CHANGELOG.md
@@ -13,6 +13,7 @@ ml).
 - [addend] #166 new action recycling and reuse
 
 ## [1.0.10-beta]
+- [bugfix] #168 can to do a trade without devices.
 
 ## [1.0.9-beta]
 - [addend] #159 external document as proof of erase of disk